@@ -15,8 +15,12 @@
     "url": "https://www.fiskercheung.com/"
   },
   "sideEffects": false,
-  "main": "dist/index.js",
-  "module": "dist/index.mjs",
+  "type": "module",
+  "exports": {
+    "import": "./dist/index.js",
+    "require": "./dist/index.cjs"
+  },
+  "main": "./dist/index.cjs",
   "files": [
     "dist"
   ],
@@ -43,7 +47,7 @@
     "release": "run-s clean lint build test format dist",
     "test": "run-p test:*",
     "test-coverage": "nyc mocha",
-    "test:js": "mocha -r esm"
+    "test:js": "mocha"
   },
   "config": {
     "commitizen": {
@@ -52,8 +56,7 @@
   },
   "mocha": {
     "bail": true,
-    "checkLeaks": true,
-    "require": "esm"
+    "checkLeaks": true
   },
   "nyc": {
     "reporter": [
@@ -63,46 +66,48 @@
   },
   "dependencies": {
     "accepts": "1.3.7",
-    "debug": "4.3.1",
+    "debug": "4.3.2",
+    "esm-utils": "1.1.0",
     "http-errors": "1.8.0",
     "lodash.template": "4.5.0",
-    "mime-types": "2.1.30",
+    "mime-types": "2.1.32",
     "parseurl": "1.3.3"
   },
   "devDependencies": {
-    "@babel/core": "7.14.2",
-    "@babel/preset-env": "7.14.2",
-    "@commitlint/cli": "12.1.4",
-    "@fisker/commitlint-config": "1.3.4",
-    "@fisker/eslint-config": "4.1.6",
-    "@fisker/eslint-config-mocha": "1.2.2",
-    "@fisker/husky-config": "4.0.4",
-    "@fisker/lint-staged-config": "3.0.4",
-    "@fisker/prettier-config": "4.0.4",
-    "@fisker/stylelint-config": "2.3.5",
-    "@rollup/plugin-commonjs": "19.0.0",
-    "@rollup/plugin-json": "4.1.0",
-    "@rollup/plugin-node-resolve": "13.0.0",
+    "@babel/core": "7.14.8",
+    "@babel/preset-env": "7.14.9",
+    "@commitlint/cli": "13.1.0",
+    "@fisker/commitlint-config": "1.3.6",
+    "@fisker/eslint-config": "7.0.9",
+    "@fisker/eslint-config-mocha": "1.2.3",
+    "@fisker/husky-config": "4.0.5",
+    "@fisker/lint-staged-config": "3.0.5",
+    "@fisker/prettier-config": "4.0.6",
+    "@fisker/stylelint-config": "2.3.6",
+    "@rollup/plugin-commonjs": "20.0.0",
+    "@rollup/plugin-node-resolve": "13.0.4",
+    "@rollup/plugin-replace": "3.0.0",
     "chai": "4.3.4",
+    "core-js": "3.16.0",
     "cz-conventional-changelog": "3.3.0",
-    "del-cli": "3.0.1",
-    "eslint": "7.26.0",
-    "esm": "3.2.25",
+    "del-cli": "4.0.1",
+    "eslint": "7.32.0",
     "get-port": "5.1.1",
-    "husky": "6.0.0",
-    "lint-staged": "11.0.0",
-    "markdownlint-cli": "0.27.1",
-    "mocha": "8.4.0",
+    "husky": "7.0.1",
+    "lint-staged": "11.1.1",
+    "markdownlint-cli": "0.28.1",
+    "mocha": "9.0.3",
     "npm-run-all": "4.1.5",
     "nyc": "15.1.0",
-    "open": "8.0.8",
-    "prettier": "2.3.0",
+    "open": "8.2.1",
+    "prettier": "2.3.2",
     "rimraf": "3.0.2",
-    "rollup": "2.48.0",
+    "rollup": "2.55.1",
     "rollup-plugin-babel": "4.4.0",
+    "rollup-plugin-prettier": "2.1.0",
     "sort-package-json": "1.50.0",
     "stylelint": "13.13.1",
-    "supertest": "6.1.3"
+    "supertest": "6.1.4"
   },
   "engines": {
     "node": ">= v6.17.1"
