@@ -1,4 +1,8 @@
 import debug from 'debug'
-import {name} from '../../package.json'
+
+import createEsmUtils from 'esm-utils'
+
+const {json} = createEsmUtils(import.meta)
+const {name} = json.loadSync('../../package.json')
 
 export default debug(name)
