@@ -1,10 +1,10 @@
 import path from 'path'
-import Connection from './connection'
-import defaults from '../defaults'
-import imports from './imports'
-import isBoolean from '../utils/is-boolean'
-import template from './template'
-import responser from './responser'
+import Connection from './connection.js'
+import defaults from '../defaults.js'
+import imports from './imports.js'
+import isBoolean from '../utils/is-boolean.js'
+import template from './template.js'
+import responser from './responser.js'
 
 class ServeDirectory {
   constructor(root, options) {
