@@ -1,4 +1,4 @@
-import isHiddenFile from './is-hidden-file'
+import isHiddenFile from './is-hidden-file.js'
 
 function notHiddenFile(file) {
   return !isHiddenFile(file)
